Move lib/posts/utils to state/posts/utils
