+++ client/blocks/reader-post-actions/index.jsx
@@ -17,7 +17,7 @@ import ReaderPostOptionsMenu from 'blocks/reader-post-options-menu';
 import { shouldShowComments } from 'blocks/comments/helper';
 import { shouldShowLikes } from 'reader/like-helper';
 import { shouldShowShare } from 'blocks/reader-share/helper';
-import { userCan } from 'lib/posts/utils';
+import { userCan } from 'state/posts/utils';
 import * as stats from 'reader/stats';
 import { localize } from 'i18n-calypso';
 import ReaderVisitLink from 'blocks/reader-visit-link';
