+++ client/post-editor/editor-confirmation-sidebar/header.jsx
@@ -10,7 +10,7 @@ import { localize } from 'i18n-calypso';
 /**
  * Internal dependencies
  */
-import * as utils from 'lib/posts/utils';
+import * as utils from 'state/posts/utils';
 
 class EditorConfirmationSidebarHeader extends PureComponent {
 	static propTypes = {
