+++ client/post-editor/editor-publish-date/index.jsx
@@ -18,7 +18,7 @@ import { intersection } from 'lodash';
  */
 import Button from 'components/button';
 import PostScheduler from './post-scheduler';
-import * as utils from 'lib/posts/utils';
+import * as utils from 'state/posts/utils';
 import { getSelectedSite } from 'state/ui/selectors';
 
 export class EditorPublishDate extends React.Component {
