+++ client/post-editor/post-editor.jsx
@@ -28,7 +28,7 @@ import SegmentedControlItem from 'components/segmented-control/item';
 import InvalidURLDialog from 'post-editor/invalid-url-dialog';
 import RestorePostDialog from 'post-editor/restore-post-dialog';
 import VerifyEmailDialog from 'components/email-verification/email-verification-dialog';
-import * as utils from 'lib/posts/utils';
+import * as utils from 'state/posts/utils';
 import EditorPreview from './editor-preview';
 import { recordStat, recordEvent } from 'state/posts/stats';
 import analytics from 'lib/analytics';
